blog/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-redis==4.10.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
hiredis==0.3.1
idna==2.8
Markdown==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.9
redis==3.0.1
requests==2.21.0
urllib3==1.24.1
